avoid crash when no file is selected

+++ mainwindow.cpp
@@ -84,10 +84,12 @@ void MainWindow::exportTriggered()
             + QDir::separator() + ConfigLoader::FILENAME;
     QString filePath = QFileDialog::getSaveFileName(this, tr("Export configuration file"),
                                                     defaultPath, tr("Json file (*.json)"));
-    if (QFile::exists(filePath)) {
-        QFile::remove(filePath);
+    if (!filePath.isEmpty()) {
+        if (QFile::exists(filePath)) {
+            QFile::remove(filePath);
+        }
+        QFile::copy(m_config->path(), filePath);
     }
-    QFile::copy(m_config->path(), filePath);
 }
 
 void MainWindow::importTriggered()
@@ -95,12 +97,14 @@ void MainWindow::importTriggered()
     QString defaultPath = QStandardPaths::writableLocation(QStandardPaths::DocumentsLocation);
     QString filePath = QFileDialog::getOpenFileName(this, tr("Import configuration file"),
                                                     defaultPath, tr("Json file (*.json)"));
-    QFile file(filePath);
-    file.open(QIODevice::ReadOnly);
-    m_config->readJson(file.readAll());
-    file.close();
-    ui->calendarWidget->setDataAdapter(m_config->data());
-    ui->calendarWidget->setFocus();
+    if (!filePath.isEmpty()) {
+        QFile file(filePath);
+        file.open(QIODevice::ReadOnly);
+        m_config->readJson(file.readAll());
+        file.close();
+        ui->calendarWidget->setDataAdapter(m_config->data());
+        ui->calendarWidget->setFocus();
+    }
 }
 
 QString MainWindow::getLongMonthName(int month)
